Institution now serves as a layer to hide the complexity of dealing with the service directly.
The institution service handles all of the institutions and holds them all in memory (for the time being).  I've also added some helper classes that could be of use in the future, specifically:

 - Event (interface)
 - ExogeneousEvent
 - Instal

The institution factory seems to be redundant at this point as all institution creation should be handled fine by the institution service.